Extract mse_value_loss
